requirements.txt
disnake
asyncio
tinydb
telethon
pillow